requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.1.0
cachetools==5.2.0
certifi==2022.12.7
charset-normalizer==2.1.1
frozenlist==1.3.3
future==0.18.2
google-api-core==2.11.0
google-api-python-client==2.69.0
google-auth==2.15.0
google-auth-httplib2==0.1.0
googleapis-common-protos==1.57.0
httplib2==0.21.0
idna==3.4
line-bot-sdk==2.3.0
multidict==6.0.3
protobuf==4.21.11
proxy-tools==0.1.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.10.2
PyJWT==2.6.0
pyobjc-core==9.0
pyobjc-framework-Cocoa==9.0
pyobjc-framework-WebKit==9.0
pyparsing==3.0.9
pywebview==3.7.2
pyzoom==0.1.15
requests==2.28.1
rsa==4.9
pyautogui==0.9.53
shortuuid==1.0.11
six==1.16.0
typing_extensions==4.4.0
uritemplate==4.1.1
urllib3==1.26.13
yarl==1.8.2
pillow==9.4.0
opencv-contrib-python==4.7.0.68
